🚚 Rename: common -> mainPage로 폴더 이동
